refactor(runner): split start job spawn
